backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
openai==1.51.0
supabase==2.9.1
langgraph==0.2.35
langchain-openai==0.2.3
mem0ai==0.1.29
python-multipart==0.0.12
httpx==0.27.2
